fix: キャッチコピーアニメーションのCLS改善及び整理
